From Brian Wheeler:
This patch gives the assembler support of '\a','\n','\r','\t', and '\\'
in string constants.

In addition, it changes (for all registers) "I reg %li is ..." to just
the value of the register.  Printing constants is also supported, but
alas, you have to specify the type (print_sc, print_ic, print_nc).


git-svn-id: https://svn.parrot.org/parrot/trunk@69 d31e2699-5ff4-0310-a27c-f18f2fbe73fe