Revisiting the problem we saw before, where tcc won't permit casting between function and data pointers.  (this is technically undefined by ANSI C).
However, it is possible to cast between them by first casting to an integer
of sufficient size.  Defined two macros for converting values betwen
function and data pointers and inserted them in the right places.

This gets things to build with tcc, and makes us more ANSI-correct.  We
might be able to remove the check for this in Configure.pl if everything
uses these macros.  (Though if JIT ever works on tcc, i will be impressed)


git-svn-id: https://svn.parrot.org/parrot/trunk@1665 d31e2699-5ff4-0310-a27c-f18f2fbe73fe