Courtesy of Andy Dougherty:
The following patch eliminates 147 warnings on my system, and puts
in place the infrastructure to get rid of even more.

Specifically, on Solaris 8 with gcc-2.8.1, and perl-5.8.0 with
64bit ints, sizeof(INTVAL) == 8, while sizeof(void *) == 4.  Hence all
casts between pointers and integers raise a warning.

The following patch expands the INTVAL2PTR and PTR2INTVAL macros in
include/parrot/parrot.h to silence the warnings in such cases.  (The
macros do not check for out-of-range errors -- they assume that the
structure of the code is such that everything is guaranteed to fit in
either a pointer or an integer.)  I also introduced a parallel set of
macros to use opcode_t (If I understand the current state of affairs,
we currently don't officially require sizeof(opcode_t) == sizeof(intval)).

I then used these macros in core.ops to eliminate a bunch of warnings.
Mostly, these are simple things like

    -  fclose((FILE *)$1);
    +  fclose(OPCODE_T2PTR(FILE *,$1));

However, in the jsr ops, where we have

    goto ADDRESS($1);

I found that if I replaced it with

    goto ADDRESS(INTVAL2PTR(opcode_t *, $1));

then core_ops_cg.c was generated with invalid code.  I think that's a
flaw in whatever generates core_ops_cg.c, but I didn't pursue that
much further.  Instead, I used a temporary variable to simplify the
expressions.  Alternatively, the generator could be changed to get the
casts right.

I can't test this too well -- parrot currently fails all tests either
with or without my patch :-).

At this point, I'm down to 74 warnings, falling into the following
general categories:

   2 cast from pointer to integer of different size
  10 cast to pointer from integer of different size

These can probably be eliminated by use of the new macros.
I just haven't gotten to them yet.  Volunteers welcome.

  48 cast increases required alignment of target type

These might each need thought.

  14 long int format, different type arg

These should go away when we do our own printf processing; in the
meantime a bandaid might be appropriate.  Volunteers welcome.


git-svn-id: https://svn.parrot.org/parrot/trunk@2035 d31e2699-5ff4-0310-a27c-f18f2fbe73fe