The following patch causes Configure to generate a 'myconfig' file that summarizes the current parrot configuration (much like perl5's 'myconfig' or perl -V output).
This patch also includes makefile targets to generate OK and Not OK
reports, and a simple tools/dev/genrpt.pl script to generate those
reports.

I think such things will be useful as we try to make parrot more portable
and assess how we are doing.

Courtesy of Andy Dougherty


git-svn-id: https://svn.parrot.org/parrot/trunk@2236 d31e2699-5ff4-0310-a27c-f18f2fbe73fe