Jonathan Sillito:
The attached patch implements a very complete set of lexical scope
semantics. For those that have been following along, here are some
highlights.

- supports (fast) access by lexical position, rather than just by name

- adds a couple more ops and some variations, including push_pad and
peek_pad

- adds a couple of keyed vtable functions for Scratchpad PMCs, so the
following work in a similar way to the various flavours of find_lex and
store_lex. The difference is that the find_lex and store_lex ops work
implicitly on the current pad, while these let you explicitly specify a pad

   new_pad P20, 0        # does not push on stack
   set P20[0;"foo"], P0  # like: store_lex Ix, Sy, Pz
   set P20[0;0], P0      # like: store_lex Ix, Iy, Pz
   set P20["foo"], P0    # like: store_lex Sx, Py
   set P20[0], P0        # like: store_lex Ix, Py
   set P3, P20[0;"foo"]  # like: find_lex Px, Iy, Sz
   set P3, P20[0;0]      # like: find_lex Px, Iy, Iz
   set P4, P20["foo"]    # like: find_lex Px, Sy
   set P4, P20[0]        # like: find_lex Px, Iy

- not based on PerlHash


git-svn-id: https://svn.parrot.org/parrot/trunk@2559 d31e2699-5ff4-0310-a27c-f18f2fbe73fe