Move PIO_isatty from macro to function
PIO_isatty is no longer a macro hardcoding a layer, but a function.
The arguments of this function also changed: They take an interpreter as
all Parrot functions and use an IO-PMC instead of a PIOHANDLE.

The helperfunctions PIO_{unix,stdio,win32}_isatty are moved to the
io_{unix,stdio,win32}.c files and are declared static.


git-svn-id: https://svn.parrot.org/parrot/trunk@4795 d31e2699-5ff4-0310-a27c-f18f2fbe73fe