#27469 [PATCH] Use named constants in iterator.pod
this patch is just a little documentation update in 'docs/iterator.pod'.
Now named constants are used for setting up the iterator in the examples.

There was even a little bug, .ITERATE_FROM_END is no longer 3, but 4.

There is now also a PIR test in 't/pmc/iter.t'.

Courtesy of Bernhard Schmalhofer


git-svn-id: https://svn.parrot.org/parrot/trunk@5560 d31e2699-5ff4-0310-a27c-f18f2fbe73fe