[perl #28093] delete_keyed of OrderedHash
After deleting a value, the OrderedHash gets confused when fetching
values that still should be stored in the hash.  Especially I got a
'PerlUndef' for the last inserted value.

The cause was that the internal PerlArray gets spliced upon deletion
of values. This messes up the mapping of the indexes stored in the
PerlHash to the values stored in the PerlArray.  The fix is not to do
a delete in the PerlArray, but to overwrite the value with a
PerlUndef.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>


git-svn-id: https://svn.parrot.org/parrot/trunk@5775 d31e2699-5ff4-0310-a27c-f18f2fbe73fe