[perl #30314] [PATCH] Use a generated makefile for languages/cola
This patch changes the languages/cola/Makefile from a manually-generated
one to one which is generated by Configure.pl from a new template.
The main thing this does is remove the explicit dependency on gcc.
Cola will now be compiled with the same compiler used to compile parrot.

With this in place, cola now builds for me.  (It still doesn't have
any tests, however . . . .)

Courtesy of Andy Dougherty <doughera@lafayette.edu>


git-svn-id: https://svn.parrot.org/parrot/trunk@6232 d31e2699-5ff4-0310-a27c-f18f2fbe73fe