[perl #30995] [PATCH] NCI for non-x86 platforms
The unmanagedstruct tried to dereference unaligned pointers.

(a) in unmanagedstruct.pmc fixed ret_int to align properly (the main
    thing was to separate INTVAL and int cases, but I added some other
    types while I was at it)
(b) config.h.in: added SHORT_SIZE and INT_SIZE to support (a)
(c) nci.t: the test assumed that (char)(3*64) would be -64 and therefore
    explicitly abs()ed the result - that the result would be 192 was not
    expected at all.
(d) pdd16_native_call.pod: added a note about signedness of the C char.
    Not that anyone ever reads documentation.

[perl #30996] [PATCH] trace prettyprinting

I added more decoding of the arguments for parrot -t, and added a little
bit of documentation for it.

[perl #31000] [PATCH] to go on top of perl #30995

To accommodate even more sizeof() combinations and to avoid duplicate


Courtesy of Jarkko Hietaniemi <jhi@iki.fi>
cases in those, ummm, cases.


git-svn-id: https://svn.parrot.org/parrot/trunk@6508 d31e2699-5ff4-0310-a27c-f18f2fbe73fe