On Solaris 8, asctime_r takes 3 parameters instead of 2.  The prototype looks like this:
char *asctime_r(const struct tm *tm, char *buf, int buflen);

This patch adds a time.c file to the solaris platform directory.  The
only change in this time.c file and the stock generic/time.c is that
the call to asctime_r passes 26 as the buflen.

The Solaris man page for asctime_r says that the result will always be
26 characters exactly.  I'm guessing the POSIX asctime assumes that
the buffer is at least 26  characters, so assuming the buffer that
Parrot_asctime_r gets is at least 26 characters isn't a new risk.

Without this patch, the latest version of parrot from CVS will not
compile on Solaris 8.

Courtesy of coneill@oneill.net


git-svn-id: https://svn.parrot.org/parrot/trunk@6589 d31e2699-5ff4-0310-a27c-f18f2fbe73fe