[perl #31930] [PATCH] Use shared libparrot in src tests, if available
This patch causes src tests to use the shared libparrot, if it exists.
This speeds them up on my system by almost a factor of 3 (due to less
copying, since the .a is so large).


[perl #31932] [PATCH] t/pmc/signal.t (more accuracy finding process to HUP)

Just a slight improvement in the logic used to find the PID of the
process to signal. (For me, it killed t/harness once, rather than the
test process.)

[ converted to a sub and used twice --leo ]

Courtesy of Jeff Clites <jclites@mac.com>


git-svn-id: https://svn.parrot.org/parrot/trunk@6811 d31e2699-5ff4-0310-a27c-f18f2fbe73fe