[perl #31938] [PATCH] add missing get_repr() op
The get_repr was present in vtable but missing as an op.
I added it and added a test in t/pmc/bigint.t

[perl #31953] [PATCH] support of multiple identifiers in a .sym

This patch supports the declaration of multiple identifiers in a
single .sym / .local directive

Probably useless for code generated by compilers.
Invaluable for hand-writen code and documentation.

Courtesy of Stephane Payrard <stef@payrard.net>


git-svn-id: https://svn.parrot.org/parrot/trunk@6813 d31e2699-5ff4-0310-a27c-f18f2fbe73fe