[perl #31966] [PATCH] memalign for Mac OS X (to allow   ARENA_DOD_FLAGS), plus a note for OpenBSD
This patch implements Parrot_memalign for Mac OS X (to allow
ARENA_DOD_FLAGS usage). It's a patch, plus a new file
(config/gen/platform/darwin/memalign.c).

Notes:
        1) In src/interpreter.c, there was a note about
Parrot_memalign_if_possible on OpenBSD having the wrong behavior. I
believe that may have been due to a bad #define in
config/gen/platform/platform_interface.h, which this patch fixes. It
would be good to see if OpenBSD behaves correctly now, after the patch
is applied.

        2) You need to enable ARENA_DOD_FLAGS in include/parrot/settings.h for
the patch to make much of a difference. (At one point it defaulted to
"on" if you had memalign, but that appears not to be the case
currently.)

Courtesy of Jeff Clites <jclites@mac.com>


git-svn-id: https://svn.parrot.org/parrot/trunk@6817 d31e2699-5ff4-0310-a27c-f18f2fbe73fe