[perl #32239] [PATCH] Parrot m4 0.0.9
this is an update of Parrot m4.

Most changes are in the documentation. There is now an explaination about
the dependency on the 'Perl Complatible Regular Expressions' library.

The macro 'printerr' is newly implemented. The macro '__file__' is also new,
but works only for a single input file.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>


git-svn-id: https://svn.parrot.org/parrot/trunk@6915 d31e2699-5ff4-0310-a27c-f18f2fbe73fe