[perl #34932]
If the user has ICU installed in a location not normally searched by
the compiler, but does specify the location of the icu-config program,
Configure.pl ought to add in the appropriate -I/path/to/icu to
ccflags, if necessary.

This patch has Configure.pl test if that additional -I flag is needed.
If ICU is found, it first tries to include <unicode/ucnv.h>.  If that
fails, then it adds -I $icuheaders to ccflags.

I don't understand ICU on Windows.  If anyone could test this patch
there, I'd appreciate it.

[As an aside:  I didn't suggest using icu-config --cppflags because
that's broken in icu-3.2 for Solaris; I've sent an appropriate patch
off to the icu list.]

Courtesy of Andy Dougherty <doughera@lafayette.edu>



git-svn-id: https://svn.parrot.org/parrot/trunk@7817 d31e2699-5ff4-0310-a27c-f18f2fbe73fe