[perl #35361] [PATCH] Skip svn revision number if none exists
In a release tarball, the .svn directories don't exist, so
config/gen/revision.pl can't find the svn revision number. That's fine,
but it's probably wiser in that case to not report "failed" during the
Configure.pl process, but to simply skip over the missing number.

Courtesy of Andy Dougherty <doughera@lafayette.edu>


git-svn-id: https://svn.parrot.org/parrot/trunk@8035 d31e2699-5ff4-0310-a27c-f18f2fbe73fe