The problem is that if you do
	loadlib "/home/rogers/foo/bar/baz"

Parrot_load_lib finds and loads "/home/rogers/foo/bar/baz.so" just fine,
but then the library isn't initialized properly because it attempts to
look up the address of "Parrot_lib_/home/rogers/foo/bar/baz_load" via
dlsym, and similarly if you add an explicit ".so" to the arg.  This
patch does the following:

   1.  Computes a trimmed "lib_name" as the part between the last "/" or
"\" and the last ".", exclusive, and uses that for constructing entry
names for dlsym.

   2.  Attempts to fix an apparent memory leak, via the addition of line
172 to src/dynext.c.  (C heavies, please scrutinize.)

   3.  Calls Parrot_locate_runtime_file on files with explicit
extensions, for completeness.

   4.  Rewrites the test for a missing extension slightly, so that it
won't get confused by a "." in a directory name.  (Not specifically
tested.)

   5.  Includes four new test cases for loadlib, with absolute/partial
pathnames with/without an extension.

   There is also this "#ifdef WIN32" business about dropping a "lib"
prefix from the name if all else fails; I am guessing that this no
longer works, but I have no way of testing it myself, so I haven't tried
to resuscitate it.  But I'm not sure that it ever worked; seems to me
that this would have to be folded into the Parrot_locate_runtime_file
library directory search in order to do the right thing.

Courtesy of Bob Rogers <rogers-perl6@rgrjr.dyndns.org>


git-svn-id: https://svn.parrot.org/parrot/trunk@8209 d31e2699-5ff4-0310-a27c-f18f2fbe73fe