Get the failing test in [while] to pass.
TclWord was stringifying the "-wrapped command as a single element list, so tcl
was trying to execute the command "{foo; bar}" instead of the commands
"foo; bar"... This is a bit of hack, and I'm sure there are more corner cases
for parsing that we'll trip over.

Also, the expected output for the test had a cut-and-paste-o.



git-svn-id: https://svn.parrot.org/parrot/trunk@10535 d31e2699-5ff4-0310-a27c-f18f2fbe73fe