[perl #38176] [PATCH] Fix string-to-integer conversion with multiple leading spaces
Until r10938, string-to-integer conversion was possible even when the
string contained more than one space before the start of the integer.

This patch modifies src/string.t to make this work again.

This patch also modifies t/op/string.t to add a test for this case, and
for other corner-cases of string-to-integer conversion.

Courtesy of Roger Browne <roger@eiffel.demon.co.uk>


git-svn-id: https://svn.parrot.org/parrot/trunk@11042 d31e2699-5ff4-0310-a27c-f18f2fbe73fe