[src] Made vtable cloning clone isa_hash member, only after global initialization has finished (that is, when isa_hash is not NULL in the source vtable).  This gives all PIR-declared classes their own isa_hash.
Also made vtable deleting delete isa_hash member, when appropriate.  This
version looks tricky, but this is to respect that PIR-declared classes have
separate isa_hash memory between their RW and RO variants.  (That's fixable in
the Class PMC, but only if you want to violate encapsulation.)

Note that this fixes the double-free error showing up in the Perl 6 fakecutable.

git-svn-id: https://svn.parrot.org/parrot/trunk@28214 d31e2699-5ff4-0310-a27c-f18f2fbe73fe