Fix TT #313, printing -0.0 on older win32 MSVCRT dlls and probably on openbsd also.
- add has_negative_zero feature, 
- fix print_n and say_n for the platforms failing the new has_negative_zero
  configure test
- enable the failing tests again

This -0.0 is a long-standing problem and finally fixes [perl #28170], 
[perl #30737], [perl #19183], [perl #60312].

Note that special yet untested platforms, like arm might also fail 
the has_negative_zero test and might fail the signbit workaround we use.
long double -0.0 will also need a workaround code like this.

There's a patch to generate the signbit() fallback at RT #30737.
Many thanks to Andy Dougherty for pointing me to the old tickets 
and for the solution which was there for 5 years.



git-svn-id: https://svn.parrot.org/parrot/trunk@36638 d31e2699-5ff4-0310-a27c-f18f2fbe73fe