[GC] Revised GC pool replenish level calculations.  The previous change in r39215 had the unfortunate property of skipping the next mark and sweep if the current mark and sweep freed up enough pool items to meet or exceed the replenish threshold.  Of course, if the next time the GC runs is because the system has run out of those pool items, there's no choice but to allocate a new arena in the pool, increasing memory usage.
This version of the algorithm avoids the skip and instead uses the replenish
level to decide whether to allocate a new arena if the sweep didn't free enough
pool items to meet the replenish level.  This results in no ballooning
allocations when one GCable type is both the constraint and frequently
recyclable and only a slight performance reduction over the
sometimes-pathological case in r39215.  (That's fixable by not immediately
populating the free list from a freshly-allocated arena.)

git-svn-id: https://svn.parrot.org/parrot/trunk@39537 d31e2699-5ff4-0310-a27c-f18f2fbe73fe