[PMC] Added a cache to speed up Class's isa_pmc.
Rather than recursing upwards for positive and negative isa tests, each Class
now has a cache of Classes it is and isn't.  This only works if the Class is
already instantiated, for obvious "don't pull the rug out from under me"
reasons.  The result is a 1.5% performance improvement on the Rakudo startup
benchmark.

git-svn-id: https://svn.parrot.org/parrot/trunk@47820 d31e2699-5ff4-0310-a27c-f18f2fbe73fe