[OO] Removed PIR overriding of Object's get_class.
You're asking for strange and wonderful and bad behavior if you do this, so
nothing does it, so why pay the NQP-rx penalty for it?  When we get a more
flexible metaobject system, people can do the kinds of things they need here in
a different (better and safer) way.

git-svn-id: https://svn.parrot.org/parrot/trunk@48785 d31e2699-5ff4-0310-a27c-f18f2fbe73fe