[GC] Optimized gc_ms_iterate_live_strings().
By moving the "don't do anything!" guard clauses from the buffer-moving
callback to before calling the callback, the stress_strings.pir benchmark runs
~6% faster and the callback gets called less than 10% of the time as before.
This appears to improve Rakudo's performance by about a percent as well.

We can live with this ugliness until someone either tries to make this function
more general (but please don't do that) or we get a better GC (so please help
with gc_massacre and the like).

Also the moral of the story is "Don't write C as if it were C++ and expect a
compiler which can't optimize across function pointer calls or source files to
generate fast code."

git-svn-id: https://svn.parrot.org/parrot/trunk@49067 d31e2699-5ff4-0310-a27c-f18f2fbe73fe