[Hash] Extracted two funcs from parrot_hash_put.
parrot_hash_get_bucket_string() is a type-safe way to find a bucket in a hash
given a STRING.

parrot_hash_store_value_in_bucket() is a reusable way to store a value in a
bucket after you've already fetched it.

Both of these functions are static.  If you ever use them outside of this file,
you will incur the wrath of premature optimization.

git-svn-id: https://svn.parrot.org/parrot/trunk@49089 d31e2699-5ff4-0310-a27c-f18f2fbe73fe