[gc] Command line option for dynamic GC threshold
The dynamic threshold roughly limits the memory wasted by objects that
could be freed but are not yet collected to a percentage of total memory
that is actually needed. Default is 25, maximum is 1000. Increasing the
dynamic threshold results in fewer GC runs and more memory consumption.

Usage:

parrot --gc-threshold=50
parrot --gc-threshold=200

git-svn-id: https://svn.parrot.org/parrot/trunk@49092 d31e2699-5ff4-0310-a27c-f18f2fbe73fe