[str] Implement partial string scan
Add a new function partial_scan to the string vtable:

INTVAL partial_scan(INTERP, STRING *src, INTVAL count, INTVAL delim);

Scans up to src->bufused bytes in src and sets src->bufused and
src->strlen to match the maximum number of complete characters found.

For variable length encodings it is possible that the buffer isn't consumed
completely. In that case, partial_scan returns the number of additional
bytes needed to decode the next incomplete characters. Otherwise returns 0.
This assumes that the initial buffer length in src->bufused is aligned
to the unit size of the string encoding.

Scans only up to count characters. Set count to -1 to disable this check.

Stops after decoding a character with code delim. Set delim to -1 to
disable this check.