[t] Fix Parrot_PMC_modulus test and add Parrot_PMC_divide
It seems that these functions take three arguments, but nothing
is done with the third argument. The function signatures of
Parrot_PMC_modulus, Parrot_PMC_divide and all similar functions
should probably be deprecated and either changed to:

1) take 2 arguments and return 1 value
2) return void and take 3 arguments

I think for unification with future design principles (such as
fixed width opcodes that return void), that option 2) should
be persued, but this issue needs to be brought up on parrot-dev.