implement PBC backrefs in constant tables
Objects previously seen in the object graphs of other constants in the same
constant table are now stored as backreferences in stead of being duplicated.
This is significant improvement for a number of common cases including, but not
limited to: namespaces, outer subs, and lexpads.

opsc.pbc, by virtue of being created using nqp, makes heavy use of all of the
above mentioned features and is now 15% smaller.