reimplement annotations for efficient load and lookup
This uses a set of sorted annotations lists of entries (one per annotation key),
which can be binary-searched to identify the active annotation. Also avoid
costly allocations by using the segment data as the storage for the entries
(similar to how bytecode does). The downside is that creating these sorted
lists may be costly.