[lib] Avoided double-declaration of a PMC param.
This should close several Coverity reports, such as CID #1287. There's still a
logical question in this code regarding the value of passing the PMC as a
parameter to these functions, then immediately retrieving it from the call
object, but that's a much larger change more worthwhile when we rewrite the PMC
to C emitter.