[GC] Simplified function for debug/non-debug cases.
The motivation for this change is to avoid a side effect in the assert,
specifically PARROT_ASSERT(--count), reported by Coverity in CID #441. Rather
than wrapping #ifdefs around the decrement, it seems clearer to extract the
normal behavior of the loop for non-debugging builds and to provide an
alternate version in the case of a debugging build.