Beginning of an M0 assembler in Perl 5
After having an M0 mindmeld with cotto++, we decided that the dynop
implementation of M0 has served us well, but development should not
continue further. It greatly helped us define the M0 spec, but the
current impedance mismatch of having to use M0 from PIR is now blocking
us from completing the spec.

This is an M0 prototype that we plan on throwing away, but which will
help us improve the spec further. There will be an assembler, which
eats M0 source code and spews out bytecode as binary data, and then
an M0 interpreter which actually executes the bytecode.

This prototype aims to have a structure much like our final M0
implementation, but implemented in Perl 5 for maximum whipupitude.

Also, the tests from this implementation will serve as the "M0 spec
tests" and will be quite useful for any implementation of M0.