[TT #1589] Remove "." from default search paths INCLUDE, LIBRARY and LANG
Apply patch from soh_cah_toa++ with small tweaks and an extra test.

Before "." was at the top of the search paths, which implied a security risk.

The other possibility from [GH #368] would be to add "." to the end of the
search paths.

Test that . is not in the library search path, and that runtime/parrot/dynext
is searched.
TODO: Deprecation, docs.