fix races in the IPv6 bind() method tests.
tests in several different scripts all try to bind to local port 1234.
this means that the suite can occasionally fail when run in
parallel, as the second script dies with an exception.

this is heavily based on the error-handling from
<t/pmc/testlib/test_server_ipv6.pir>.  it increments the port number a
few times to try and find one it can use.  if for some reason it can't,
it bails out with the original exception.