Wrap PAST::Compiler in NQP
Most of the methods are single large Q:PIR blocks, although I had to
convert most of inline() because it would munge the '%r' and '%t'
string literals with registers.  This would seem to indicate that
inline needs to be smarter about replacing those literals, but we'll
get to that another day.